Use floating-point numbers in the linear problems.
Upgrade Farkle version.

+++ s4/linear-and-network-programming/LinearProgrammingProblemParser/LPTypes.fs
@@ -1,6 +1,6 @@
 namespace LinearProgrammingProblemParser.DomainTypes
 
-type Number = int
+type Number = float
 
 type X = X of int
